Added a 'RandomWalk'ing ability (through a class)
Added a Walk abstract class, every class of which will be a different type of 'walk' of the game board. Also added a concrete 'RandomWalk', which randomly makes moves on the board in hopes of chancing upon the solved board. Also cleaned up the drivers/added a new one for random walks and modified the makefile to reflect this