Changed the way that possible moves are found in SlidingBrickPuzzle
The previous functionality was broken when checking to the right and down (for pieces that are larger than a single square). Removed the convoluted check_direction function to fix this and split the functionality into four different functions, one for each direction (check_up/down/left/right). Though it ends up feeling like there is repeated code (and there sort of is), the logic is much more clear this way.