Added all of the classes for uninformed searches.
Added a generic UninformedWalk class for using uninformed search strategies. BreadthFirstWalk and DepthFirstWalk extend it simply by changing the open list's data structure (queue for the former, stack the latter). (See the repositories for Stack and Queue for the specific implementations used here.) Also added the basics of an IterativeDeependingWalk class, but it doesn't quite work yet.