Added a slightly incorrect heuristic function.
The heuristic is 'slightly incorrect' in the sense that it does nto compute the correct values for goals which are on the bottom or right borders (at the very least; it's possible there are more problem situations of which I am not thinking). I've added the prototype for a function to fix this but have not implemented it yet. Also added testing to the correct.cpp driver.